backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
azure-search-documents